[#166734445] mock redis utils for tests
